población de datos exitosa

+++ database/migrations/2024_08_24_054238_create_products_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('fabricante');
             $table->text('descripcion');
             $table->decimal('precio', 10, 2);
-            $table->decimal('descuento');
+            $table->integer('descuento');
             $table->integer('stock');
             $table->timestamp('fecha_agregada')->useCurrent();
             //Foreign key
